requirements.txt
asgiref==3.5.0                         
certifi==2022.5.18.1                   
charset-normalizer==2.0.12             
coreapi==2.3.3                         
coreschema==0.0.4                      
Django==4.0.2                          
django-cors-headers==3.11.0            
djangorestframework==3.13.1            
djangorestframework-simplejwt==5.2.2   
drf-extra-fields==3.4.0                
drf-yasg==1.20.0                       
filetype==1.1.0                        
idna==3.3                              
inflection==0.5.1                      
itypes==1.2.0                          
Jinja2==3.1.2                          
MarkupSafe==2.1.1                      
packaging==21.3                        
Pillow==9.2.0                          
psycopg2==2.9.5                        
psycopg2-binary==2.9.5                 
psycopg2-pool==1.1                     
PyJWT==2.6.0                           
pyparsing==3.0.9                       
pytz==2021.3                           
requests==2.27.1                       
ruamel.yaml==0.17.21                   
ruamel.yaml.clib==0.2.6                
sqlparse==0.4.2                        
tzdata==2021.5                         
uritemplate==4.1.1                     
urllib3==1.26.9                        
